:art: restructure drops  endpoints
