+++ controllers/drop_controller.js
@@ -1,0 +1,9 @@
+const dropDAO = require("../daos/drop_dao");
+
+const dropCards = async function () {
+  return await dropDAO.getTemplates();
+};
+
+module.exports = {
+  dropCards,
+};
