+++ use_cases/drop_interactor.js
@@ -1,0 +1,7 @@
+const executeDropCards = async function (dropController) {
+  return await dropController.dropCards();
+};
+
+module.exports = {
+  executeDropCards,
+};
